requirements.txt
asgiref==3.2.10
django-crispy-forms
pytz==2020.1
sqlparse==0.3.1
Django==3.2
Pillow==8.1.2
pip3 install django-recaptcha && gunicorn